fix startup with meteor command
This fixes starting meteor with the Can't listen on port 3000. Perhaps another Meteor is running?

Running two copies of Meteor in the same application directory
will not work. If something else is using port 3000, you can
specify an alternative port with --port <port>. command. Before, there was no check to see if a settings file was added, causing the server-side startup to break. Without a settings file, the server would crash. Now, the server still starts regardless of a settings file.