Swap Meteor.settings private variable
On line 3, we need to *set* the private Meteor MAIL_URL variable to whatever is stored in our local environment. We store keys inside our .bashrc or .zshrc file so those important keys, tokens, etc aren't exposed to Git. When the app initializes and hits this file, we store the value of MAIL_URL from our local machine to the Meteor instance running. Sorry about the extra work regarding this!