use of alanning:roles
When generating the accounts on the first launch, the administrators
account has an ‘admin’ role set.
This allows them to access a group of routes name adminRoutes. The
redirection is based on the existing redirectHandler.
The admin nav is only shown if the user is in role admin.
It’s intended to be the start of an administrator panel (settings,
users, etc.)